Core-service: add cud operation for LIKEs

+++ core-server/src/main/java/com/imagespace/core/domain/entity/Post.java
@@ -24,7 +24,7 @@ public class Post extends BaseEntity {
 
     @Column UUID source;
 
-    @Column int likes;
+    @Column long likes;
 
     @ManyToOne
     @JoinColumn(name = "account_id")
